+++ src/main/java/br/com/padariaweb/service/ILoginService.java
@@ -2,7 +2,7 @@ package br.com.padariaweb.service;
 
 import br.com.padariaweb.entity.Funcionario;
 
-public interface ILoginService extends IGenericoCRUDManager<Funcionario, Integer>{
+public interface ILoginService extends IGenericoCRUDManager<Funcionario, Long>{
 
 	public Funcionario verificaFuncionario(Funcionario func);
 }
